% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\name{answer_delRow_equal}
\alias{answer_delRow_equal}
\title{在记录结果中按答案删除精确匹配数据}
\usage{
answer_delRow_equal(data, answer_filter_equal = "?")
}
\arguments{
\item{data}{数据}

\item{answer_filter_equal}{问题精确匹配}
}
\value{
返回值
}
\description{
在记录结果中按答案删除精确匹配数据
}
\examples{
answer_delRow_equal();
}
