% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealLog.R
\name{dealLog}
\alias{dealLog}
\title{读取日志文件并进行处理，主函数}
\usage{
dealLog(file = "data-raw/input/2019-07-17.txt")
}
\arguments{
\item{file}{日志文件}
}
\value{
返回数据框
}
\description{
读取日志文件并进行处理，主函数
}
\examples{
dealLog();
针对数据进行删除处理，不包括------
}
