% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealLog.R
\name{dealLog2}
\alias{dealLog2}
\title{用于数据处理}
\usage{
dealLog2(file = "data-raw/input/2019-07-26.txt",
  file_filter_equal = c("好"), file_filter_like = c("http://"),
  question_filter_equal = c("在吗"),
  question_filter_like = c("在的"),
  answer_filter_equal = c("不客气"),
  answer_filter_like = c("不用谢"), row_sep = "")
}
\arguments{
\item{file}{日志名}

\item{file_filter_equal}{文件级精确匹配}

\item{file_filter_like}{文件级近似匹配}

\item{question_filter_equal}{问题精确匹配}

\item{question_filter_like}{问题近似匹配}

\item{answer_filter_equal}{答案精确匹配}

\item{answer_filter_like}{答案近似匹配}

\item{row_sep}{分隔符}
}
\value{
返回值
}
\description{
用于数据处理
}
\examples{
data <-dealLog2(file='data-raw/input/2019-07-26.txt',
file_filter_equal,
file_filter_like,
question_filter_equal,
question_filter_like,
answer_filter_equal,
answer_filter_like,
row_sep );
}
