% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealLog.R
\name{dealLog_dir}
\alias{dealLog_dir}
\title{批量处理日志文件}
\usage{
dealLog_dir(files = "data-raw/input", file_filter_equal = c("好"),
  file_filter_like = c("http://"), question_filter_equal = c("在吗"),
  question_filter_like = c("在的"),
  answer_filter_equal = c("不客气"),
  answer_filter_like = c("不用谢"), row_sep = "")
}
\arguments{
\item{files}{日志目录}

\item{file_filter_equal}{文件过滤器}

\item{file_filter_like}{文件过滤器}

\item{question_filter_equal}{问题过滤器}

\item{question_filter_like}{问题过滤器}

\item{answer_filter_equal}{答案过滤器}

\item{answer_filter_like}{答案过滤器}

\item{row_sep}{分隔符}
}
\value{
返回值
}
\description{
批量处理日志文件
}
\examples{
dealLog_dir();
}
