% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\name{log_combine_gp}
\alias{log_combine_gp}
\title{根据Id合并数据处理}
\usage{
log_combine_gp(data, sep = "")
}
\arguments{
\item{data}{单项数据}

\item{sep}{多行分隔符}
}
\value{
返回值
}
\description{
根据Id合并数据处理
}
\examples{
log_conbine_gp();
}
