% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLog.R
\name{log_delAutLong30}
\alias{log_delAutLong30}
\title{删除aut_id长于30的字符}
\usage{
log_delAutLong30(data)
}
\arguments{
\item{data}{log数据}
}
\value{
返回值
}
\description{
删除aut_id长于30的字符
}
\examples{
log_delAutLong30();
}
