% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLog.R
\name{log_delContentNull}
\alias{log_delContentNull}
\title{针对数据进行处理，删除内容中的""}
\usage{
log_delContentNull(data)
}
\arguments{
\item{data}{数据}
}
\value{
返回值
}
\description{
针对数据进行处理，删除内容中的""
}
\examples{
log_delContentNull();
}
