% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLog.R
\name{log_delLeftSpace}
\alias{log_delLeftSpace}
\title{删除数据中的左空格}
\usage{
log_delLeftSpace(data, space_count = 1)
}
\arguments{
\item{data}{log数据}

\item{space_count}{空格数}
}
\value{
返回值
}
\description{
删除数据中的左空格
}
\examples{
log_delLeftSpace();
}
