% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLog.R
\name{log_delRow_contains}
\alias{log_delRow_contains}
\title{删除日志中包含多个关键词的行}
\usage{
log_delRow_contains(data, keys)
}
\arguments{
\item{data}{日志数据}

\item{keys}{多个关键词}
}
\value{
返回数据
}
\description{
删除日志中包含多个关键词的行
}
\examples{
log_delRow_contains();
}
