% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLog.R
\name{log_delRow_equals}
\alias{log_delRow_equals}
\title{删除数据中完全匹配keys的行记录}
\usage{
log_delRow_equals(data, keys)
}
\arguments{
\item{data}{数据}

\item{keys}{关键词}
}
\value{
返回值
}
\description{
删除数据中完全匹配keys的行记录
}
\examples{
log_delRow_equals();
}
