% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\name{question_delRow_contains}
\alias{question_delRow_contains}
\title{针对问题清单进行模糊匹配}
\usage{
question_delRow_contains(data, question_filter_like)
}
\arguments{
\item{data}{数据}

\item{question_filter_like}{过滤器}
}
\value{
返回值
}
\description{
针对问题清单进行模糊匹配
}
\examples{
question_delRow_contains();
}
