% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\name{question_delRow_equal}
\alias{question_delRow_equal}
\title{在记录结果中按问题删除精确匹配数据}
\usage{
question_delRow_equal(data, question_filter_equal = "?")
}
\arguments{
\item{data}{数据}

\item{question_filter_equal}{问题精确匹配}
}
\value{
返回值
}
\description{
在记录结果中按问题删除精确匹配数据
}
\examples{
question_delRow_equal();
}
