% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLog.R
\name{read_kflog}
\alias{read_kflog}
\title{读取网商的订单日志信息}
\usage{
read_kflog(file = "data-raw/input/2019-07-17.txt")
}
\arguments{
\item{file}{文件名,要求txt另存为utf-8格式}
}
\value{
返回值
}
\description{
读取网商的订单日志信息
}
\examples{
read_kflog('data-raw/input/2019-07-17.txt');
}
