% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLog.R
\name{read_kflog2}
\alias{read_kflog2}
\title{读取日志文件}
\usage{
read_kflog2(file = "data-raw/input/2019-07-17.txt")
}
\arguments{
\item{file}{文件名}
}
\value{
返回值
}
\description{
读取日志文件
}
\examples{
read_kflog2();
针对数据进行删除处理，不包括------
}
