% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLog.R
\name{read_kflog2_new}
\alias{read_kflog2_new}
\title{读取日志并进行格式化处理}
\usage{
read_kflog2_new(file = "data-raw/2020-05-03～2020-05-03-final.txt")
}
\arguments{
\item{file}{文件名}
}
\value{
返回值
}
\description{
读取日志并进行格式化处理
}
\examples{
read_log2_new()
}
