% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLog.R
\name{read_kflog_new}
\alias{read_kflog_new}
\title{读取新的日志信息}
\usage{
read_kflog_new(file = "data-raw/2020-05-03～2020-05-03-final.txt")
}
\arguments{
\item{file}{日志文件}
}
\value{
返回数据框
}
\description{
读取新的日志信息
}
\examples{
read_kflog_new()
}
